pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "phase-forensics"
version = "0.1.0"
description = "Training-Free Synthetic Image Detection via Local Phase Coherence Violation in the Wavelet Domain"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "dtcwt>=0.12.0",
    "PyWavelets>=1.4.0",
    "Pillow>=9.5.0",
    "opencv-python>=4.7.0",
    "scikit-image>=0.20.0",
    "scikit-learn>=1.2.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "tqdm>=4.65.0",
    "PyYAML>=6.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.setuptools.packages.find]
include = ["src*"]
